Make timing count return a u64 and handle rollovers internally.
This means timing_count should be called every ~4mins at least to detect the rollover.

+++ src/swift_nap_io.h
@@ -102,7 +102,7 @@ u32 last_exti_count();
 void wait_for_exti();
 
 //void timing_strobe_setup();
-u32 timing_count();
+u64 timing_count();
 u32 timing_count_latched();
 void timing_strobe(u32 falling_edge_count);
